Enhanced codebase through refactoring in aprguniqueproblems module

+++ AprgUniqueProblems/AprgUniqueProblems/src/AprgUniqueProblems/CombiningAlgorithms/NearestEqualCells.cpp
@@ -25,40 +25,36 @@ NearestEqualCells::CoordinatePair NearestEqualCells::getNearestEqualPair(Value c
     // the algorithm is only O(n*sqrt(n)).
 
     Coordinates coordinatesWithValue(getCoordinatesWithThisValue(value));
-    if (coordinatesWithValue.size() <= static_cast<unsigned int>(sqrt(m_valueMatrix.getNumberOfCells()))) {
+    if (static_cast<int>(coordinatesWithValue.size()) <= static_cast<int>(sqrt(m_valueMatrix.getNumberOfCells()))) {
         return getNearestEqualPairByCheckingAllPairs(value);
     } else {
-        return getNearestEqualPairUsingBfs(value);
-    }
+        return getNearestEqualPairUsingBfs(value);    }
 }
 
-NearestEqualCells::CoordinatePair NearestEqualCells::getNearestEqualPairByCheckingAllPairs(Value const value) const {
-    // Algorithm 1: Go through all pairs of cells with letter c, and calculate the minimum distance between such cells.
+NearestEqualCells::CoordinatePair NearestEqualCells::getNearestEqualPairByCheckingAllPairs(Value const value) const {    // Algorithm 1: Go through all pairs of cells with letter c, and calculate the minimum distance between such cells.
     // This will take O(k2) time where k is the number of cells with letter c.
 
     CoordinatePair result{};
     Coordinates coordinates(getCoordinatesWithThisValue(value));
     auto twoCoordinatesCombinations(
-        CombinationsGeneration<Coordinates>::generateCombinationsWithLength(coordinates, 2U));
+        CombinationsGeneration<Coordinates>::generateCombinationsWithLength(coordinates, 2));
 
     bool isFirst(true);
-    unsigned int minimumDistance{};
+    int minimumDistance{};
     for (auto const& twoCoordinates : twoCoordinatesCombinations) {
-        if (twoCoordinates.size() >= 2U) {
+        if (twoCoordinates.size() >= 2) {
             if (isFirst) {
                 minimumDistance = getDistance(twoCoordinates.at(0), twoCoordinates.at(1));
                 result = {twoCoordinates.at(0), twoCoordinates.at(1)};
                 isFirst = false;
             } else {
-                unsigned int distance = getDistance(twoCoordinates.at(0), twoCoordinates.at(1));
+                int distance = getDistance(twoCoordinates.at(0), twoCoordinates.at(1));
                 if (distance < minimumDistance) {
                     minimumDistance = distance;
-                    result = {twoCoordinates.at(0), twoCoordinates.at(1)};
-                }
+                    result = {twoCoordinates.at(0), twoCoordinates.at(1)};                }
             }
         }
-    }
-    return result;
+    }    return result;
 }
 
 NearestEqualCells::CoordinatePair NearestEqualCells::getNearestEqualPairUsingBfs(Value const value) const {
@@ -70,26 +66,22 @@ NearestEqualCells::CoordinatePair NearestEqualCells::getNearestEqualPairUsingBfs
 }
 
 void NearestEqualCells::initializeGraph() {
-    m_valueMatrix.iterateAllThroughYAndThenX([&](unsigned int const x, unsigned int const y) {
+    m_valueMatrix.iterateAllThroughYAndThenX([&](int const x, int const y) {
         if (m_valueMatrix.isInside(x + 1, y)) {
             m_coordinateGraph.connect(Coordinate(x, y), Coordinate(x + 1, y));
-        }
-        if (m_valueMatrix.isInside(x, y + 1)) {
+        }        if (m_valueMatrix.isInside(x, y + 1)) {
             m_coordinateGraph.connect(Coordinate(x, y), Coordinate(x, y + 1));
         }
-    });
-}
+    });}
 
 NearestEqualCells::Coordinates NearestEqualCells::getCoordinatesWithThisValue(Value const value) const {
     Coordinates result;
-    m_valueMatrix.iterateAllThroughYAndThenX([&](unsigned int const x, unsigned int const y) {
+    m_valueMatrix.iterateAllThroughYAndThenX([&](int const x, int const y) {
         if (m_valueMatrix.getEntry(x, y) == value) {
             result.emplace_back(x, y);
-        }
-    });
+        }    });
     return result;
 }
-
 NearestEqualCells::Coordinate NearestEqualCells::getFirstCoordinateUsingBfs(Value const value) const {
     Coordinates coordinates(getCoordinatesWithThisValue(value));
     Bfs bfs(m_coordinateGraph, coordinates);
@@ -107,27 +99,24 @@ NearestEqualCells::Coordinate NearestEqualCells::getCoordinateUsingBfs(
     Value const value, Coordinate const& firstCoordinate, Bfs& bfs) const {
     Coordinate result{};
     bool isFirst(true);
-    unsigned int minimumDistance{};
+    int minimumDistance{};
     for (auto const& vertexDistancePair : bfs.getEndVertexToDistanceCountMap()) {
         Coordinate const& coordinate(vertexDistancePair.first);
-        unsigned int distance(vertexDistancePair.second);
+        int distance(vertexDistancePair.second);
         if (value == m_valueMatrix.getEntry(coordinate.first, coordinate.second) && firstCoordinate != coordinate) {
             if (isFirst) {
-                minimumDistance = distance;
-                result = coordinate;
+                minimumDistance = distance;                result = coordinate;
                 isFirst = false;
             } else if (distance < minimumDistance) {
-                minimumDistance = distance;
-                result = coordinate;
+                minimumDistance = distance;                result = coordinate;
             }
         }
     }
     return result;
 }
 
-unsigned int NearestEqualCells::getDistance(Coordinate const& coordinate1, Coordinate const& coordinate2) const {
+int NearestEqualCells::getDistance(Coordinate const& coordinate1, Coordinate const& coordinate2) const {
     return getPositiveDelta(coordinate1.first, coordinate2.first) +
            getPositiveDelta(coordinate1.second, coordinate2.second);
 }
-
 }  // namespace alba
