+++ AprgUniqueProblems/AprgUniqueProblems/src/AprgUniqueProblems/DynamicProgramming/Other/ElevatorWeightProblem.cpp
@@ -9,16 +9,14 @@ ElevatorWeightProblem::ElevatorWeightProblem(Weight const maximumElevatorWeight,
       m_peopleWeights(peopleWeights),
       m_numberOfRidesAndLastWeights(getNumberOfPeopleSubsets(), NumberOfRidesAndWeight{}) {}
 
-unsigned int ElevatorWeightProblem::getNumberOfOptimalRides() {
-    unsigned int result(0);
+int ElevatorWeightProblem::getNumberOfOptimalRides() {
+    int result(0);
     for (PeopleBits peopleBits = 1; peopleBits < getNumberOfPeopleSubsets(); peopleBits++) {
         // initial value: n+1 rides are needed
-        m_numberOfRidesAndLastWeights[peopleBits] = NumberOfRidesAndWeight{getNumberOfPeople() + 1, 0};
-        for (Person person = 0; person < getNumberOfPeople(); person++) {
+        m_numberOfRidesAndLastWeights[peopleBits] = NumberOfRidesAndWeight{getNumberOfPeople() + 1, 0};        for (Person person = 0; person < getNumberOfPeople(); person++) {
             if (isPersonIncluded(peopleBits, person)) {
                 NumberOfRidesAndWeight newOption = m_numberOfRidesAndLastWeights[removePerson(peopleBits, person)];
-                if (newOption.second + m_peopleWeights.at(person) <= m_maximumElevatorWeight) {
-                    // add p to an existing ride
+                if (newOption.second + m_peopleWeights.at(person) <= m_maximumElevatorWeight) {                    // add p to an existing ride
                     newOption.second += m_peopleWeights.at(person);
                 } else {
                     // reserve a new ride for p
@@ -30,16 +28,14 @@ unsigned int ElevatorWeightProblem::getNumberOfOptimalRides() {
             }
         }
     }
-    unsigned int allPeopleBits = getNumberOfPeopleSubsets() - 1;
-    if (allPeopleBits < m_numberOfRidesAndLastWeights.size()) {
+    int allPeopleBits = getNumberOfPeopleSubsets() - 1;
+    if (allPeopleBits < static_cast<int>(m_numberOfRidesAndLastWeights.size())) {
         result = m_numberOfRidesAndLastWeights.at(allPeopleBits).first + 1;
     }
-    return result;
-}
+    return result;}
 
 bool ElevatorWeightProblem::isPersonIncluded(PeopleBits const peopleBits, Person const person) const {
-    return peopleBits & getProductBits(person);
-}
+    return peopleBits & getProductBits(person);}
 
 ElevatorWeightProblem::Person ElevatorWeightProblem::getNumberOfPeople() const { return m_peopleWeights.size(); }
